authorization with policies

+++ app/Http/Controllers/ThreadsController.php
@@ -70,7 +70,8 @@ class ThreadsController extends Controller
 
     public function destroy($channel,Thread $thread)
     {
-//        $thread->replies()->delete();
+
+        $this->authorize('update',$thread);
         $thread->delete();
         if(request()->wantsJson())
         {
